requirements.txt
asgiref==3.2.3
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-bleach==0.6.1
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
idna==2.9
mongoengine==0.19.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pymongo==3.10.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==3.2.0
social-auth-storage-mongoengine==1.0.1
sqlparse==0.3.0
urllib3==1.25.8
webencodings==0.5.1
